fix: get api 한번만 호출하도록 변경
